[CSPM][SEC-5583] migrate compliance check to fx (#14052)
* move compliance check to fx

* move stuff around so that it's importable

* rename file

* migrate cluster agent to use new component enabled check

* bug fixes

* improve cluster agent config setup

* fix windows build
