+++ cmd/cluster-agent/app/compliance_cmd.go
@@ -9,13 +9,10 @@
 package app
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/common"
-	"github.com/DataDog/datadog-agent/cmd/security-agent/app"
-	"github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/cmd/security-agent/app/subcommands/check"
+	"github.com/DataDog/datadog-agent/comp/core"
 )
 
 var (
@@ -26,17 +23,11 @@ var (
 )
 
 func init() {
-	checkCmd := app.CheckCmd()
-	checkCmd.PreRunE = func(cmd *cobra.Command, args []string) error {
-		// we'll search for a config file named `datadog-cluster.yaml`
-		config.Datadog.SetConfigName("datadog-cluster")
-		err := common.SetupConfig(confPath)
-		if err != nil {
-			return fmt.Errorf("unable to set up global cluster agent configuration: %w", err)
-		}
-		return nil
-	}
+	bundleParams := core.BundleParams{
+		ConfFilePath: confPath,
+		ConfigName:   "datadog-cluster",
+	}.LogForOneShot(string(loggerName), "off", true)
 
-	complianceCmd.AddCommand(checkCmd)
+	complianceCmd.AddCommand(check.Commands(bundleParams)...)
 	ClusterAgentCmd.AddCommand(complianceCmd)
 }
