+++ cmd/security-agent/app/subcommands/check/command_unsupported.go
@@ -5,7 +5,7 @@
 
 //go:build windows || !kubeapiserver
 
-package app
+package check
 
 import (
 	"github.com/spf13/cobra"
@@ -13,6 +13,6 @@ import (
 	"github.com/DataDog/datadog-agent/cmd/security-agent/app/common"
 )
 
-func CheckCommands(globalParams *common.GlobalParams) []*cobra.Command {
+func SecAgentCommands(globalParams *common.GlobalParams) []*cobra.Command {
 	return nil
 }
